# [Performance] added `SpanAttributes.Add(string,object)` method for deferred `string`-ification