## Changes

This is a performance fix aimed at making it possible to defer `string`-ification of complex types until the `Activity` is being processed by the OTel pipeline.

This type of performance work is possible when working directly on `System.Diagnostics.ActivityTagsCollection`, but not when working with `OpenTelemetry.Api.Trace.SpanAttributes`.

## Merge requirement checklist

* [x] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-dotnet/blob/main/CONTRIBUTING.md) guidelines followed (nullable enabled, static analysis, etc.)
* [x] Unit tests added/updated
* [ ] Appropriate `CHANGELOG.md` files updated for non-trivial changes
* [x] Changes in public API reviewed (if applicable)