@Aaronontheweb 2023-07-26
Since multiple comments have asked for benchmarks.

While working on our commercial library built on top of the OTEL API we conducted extensive performance testing, and "deferred stringification" was the most significant performance impact we made. Full post is here: https://petabridge.com/blog/phobos-2.4-opentelemetry-optimizations/

## Original Code

Simplified this some, but this is what our span creation looks like around the hot path
```csharp
private TelemetrySpan CreateActiveSpan(object underlyingMsg, SpanContext? context = null,
    Baggage baggage = default, bool followFromActiveSpan = true, long? startTimeUtcTicks = null)
{
    var startTime = startTimeUtcTicks != null
        ? new DateTimeOffset(startTimeUtcTicks.Value, TimeSpan.Zero)
        : DateTimeOffset.UtcNow;

    TelemetrySpan sp = null;
    if (context == null)
    {
        sp = Tracer.StartRootSpan(underlyingMsg.GetType().GetOperationName(),
            SpanKind.Consumer, InitialAttributes, startTime: startTime);
            
        Tracer.WithSpan(sp); // mark the current span as active
    }
    else
    {
        sp = Tracer.StartActiveSpan(underlyingMsg.GetType().GetOperationName(),
            SpanKind.Consumer, parentContext: context.Value, InitialAttributes, startTime: startTime);
    }

    if (startTimeUtcTicks != null)
    {
        sp.AddEvent("waiting", startTime);
    }
    var attributes = new SpanAttributes();
    attributes.Add("content", underlyingMsg.ToString());
    sp.AddEvent("message", attributes);

    return sp;
}
```

### Original benchmark values

```

BenchmarkDotNet v0.13.6, Windows 11 (10.0.22621.1992/22H2/2022Update/SunValley2)
12th Gen Intel Core i7-1260P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 7.0.302
  [Host]     : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2


```
|                     Method | MessageKind |     Mean |   Error |  StdDev |      Req/sec |   Gen0 | Allocated |
|--------------------------- |------------ |---------:|--------:|--------:|-------------:|-------:|----------:|
|          **CreateNewRootSpan** |   **Primitive** | **119.1 ns** | **2.17 ns** | **2.03 ns** | **8,399,591.08** | **0.0246** |     **232 B** |
|            CreateChildSpan |   Primitive | 123.8 ns | 2.47 ns | 2.42 ns | 8,076,859.69 | 0.0246 |     232 B |
| CreateChildSpanWithBaggage |   Primitive | 122.3 ns | 2.18 ns | 2.04 ns | 8,179,444.13 | 0.0246 |     232 B |
|          **CreateNewRootSpan** |       **Class** | **164.4 ns** | **2.80 ns** | **2.62 ns** | **6,081,681.47** | **0.0348** |     **328 B** |
|            CreateChildSpan |       Class | 170.2 ns | 1.94 ns | 1.72 ns | 5,873,753.13 | 0.0348 |     328 B |
| CreateChildSpanWithBaggage |       Class | 169.6 ns | 3.17 ns | 2.97 ns | 5,896,706.30 | 0.0348 |     328 B |
|          **CreateNewRootSpan** |      **Record** | **222.3 ns** | **4.48 ns** | **5.50 ns** | **4,497,668.25** | **0.0730** |     **688 B** |
|            CreateChildSpan |      Record | 222.9 ns | 4.27 ns | 4.75 ns | 4,485,591.41 | 0.0730 |     688 B |
| CreateChildSpanWithBaggage |      Record | 226.7 ns | 2.80 ns | 2.48 ns | 4,411,065.79 | 0.0730 |     688 B |


As you can see - both the memory allocation in the hot path AND the end to end latency changes with the _type of message being processed_:

* Primitive is just a `string` - and its `object.ToString()` call is a no-op;
* Class is a simple `class` that uses an overridden `object.ToString()` implementation that uses string interpolation to print out all of the field values.
* Record is a `record` implementation of the same structure as the class, but uses `record`'s built-in `ToString()` method which is rather expensive (relies on `StringBuilder`s et al internally.)

## New code

This code uses the existing public OTEL APIs - I allocate an additional `IEnumerable<KeyValuePair<string, object>>` and pass these into the CTOR of `SpanAttributes`. However, there's at least two areas where we pass in `InitialSpanAttributes` to a `TelemeterySpan` where adding a `AddAttribute(string, object)` would benefit even more. Plus adding that method allows me to eliminate an additional enumerator:

```csharp
// rest of  CreateActiveSpan method
static IEnumerable<KeyValuePair<string, object>> CreateSpanAttributes(object msg)
{
    yield return new KeyValuePair<string, object>("content", msg);
}

var attributes = new SpanAttributes(CreateSpanAttributes(underlyingMsg));
sp.AddEvent("message", attributes);

return sp;
```

### Updated benchmark values

```

BenchmarkDotNet v0.13.6, Windows 11 (10.0.22621.1992/22H2/2022Update/SunValley2)
12th Gen Intel Core i7-1260P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 7.0.302
  [Host]     : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2


```
|                     Method | MessageKind |     Mean |   Error |  StdDev |      Req/sec |   Gen0 | Allocated |
|--------------------------- |------------ |---------:|--------:|--------:|-------------:|-------:|----------:|
|          **CreateNewRootSpan** |   **Primitive** | **152.0 ns** | **2.58 ns** | **2.42 ns** | **6,580,292.48** | **0.0322** |     **304 B** |
|            CreateChildSpan |   Primitive | 149.7 ns | 1.98 ns | 1.85 ns | 6,681,595.57 | 0.0322 |     304 B |
| CreateChildSpanWithBaggage |   Primitive | 150.3 ns | 2.71 ns | 2.53 ns | 6,653,843.06 | 0.0322 |     304 B |
|          **CreateNewRootSpan** |       **Class** | **150.1 ns** | **2.74 ns** | **2.29 ns** | **6,660,638.95** | **0.0322** |     **304 B** |
|            CreateChildSpan |       Class | 149.1 ns | 2.38 ns | 2.11 ns | 6,704,915.73 | 0.0322 |     304 B |
| CreateChildSpanWithBaggage |       Class | 151.2 ns | 2.87 ns | 2.69 ns | 6,613,827.37 | 0.0322 |     304 B |
|          **CreateNewRootSpan** |      **Record** | **147.8 ns** | **2.80 ns** | **2.62 ns** | **6,766,604.10** | **0.0322** |     **304 B** |
|            CreateChildSpan |      Record | 147.0 ns | 2.75 ns | 2.57 ns | 6,801,670.75 | 0.0322 |     304 B |
| CreateChildSpanWithBaggage |      Record | 149.7 ns | 2.87 ns | 2.95 ns | 6,678,899.59 | 0.0322 |     304 B |

Default memory allocations go up for the primitive case, but otherwise the hot-path's memory consumption and latency is now consistent regardless of the input data type - this is because the expensive operation of rendering the string now happens out-of-band inside the OTEL processing / export pipeline instead.

If you'd like me to add a version of this benchmark to this PR, I'd be happy to do it - but these are our numbers demonstrating that the _techique_ of deferring `string`-ification is conducive to better performance. Adding the overload I've included in this PR will make it easier to do that.